Non_Divisible_Subset_Statement
Problem Statement:
Given a set of distinct integers, print the size of a maximal subset S of where the sum of any 2 numbers in S' is not evenly divisible by 2.
For example, the array S= [19,10,12,10,24,25,22] and k=4. One of the arrays that can be created is S'[0] = [10,12,25]. Another is S'[1] = [19,22,24]. After testing all permutations, the maximum length solution array has 3 elements. 

Function Description:
Complete the nonDivisibleSubset function in the editor below. It should return an integer representing the length of the longest subset of S meeting the criteria.
nonDivisibleSubset has the following parameter(s):
    S: an array of integers
    k: an integer

For details, refer
https://www.hackerrank.com/challenges/non-divisible-subset/problem